requirements.txt
pyserial #python3-pip

